copbet_py/requirements.txt
numpy>=1.21
scipy>=1.7
scikit-learn>=1.0
matplotlib>=3.4
networkx>=2.6
bctpy>=0.5
